Players in lobbies and in game are unified as wormhole.Player.  Protocol file for players complete.
There is a regression in the code for, possibly due to this change, more
likely from some other patch.  Ownership of a BaseObject does not
properly transfer.  Production of ships changes to appropriate player,
however, defenses do not change over, and destroy as many new ships as
is appropriate for the defense strength.